agentos-pessoas/requirements.txt
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.20.0,<0.25.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
loguru>=0.7.0,<0.8.0
httpx[http2]>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
motor>=3.3.2